requirements.txt
jira>=3.5.0
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
